refactor: Transition to modular architecture and clean up dead code
Decompose App.jsx into pure composition structure (61% LoC reduction)

Extract UI components (Header, StatusDashboard, LogHistory, etc.)

Isolate business logic into useInnerOrbit custom hook

Remove unused assets and CSS files for project cleanup
