+++ src/hooks/useInnerOrbit.js
@@ -1,0 +1,157 @@
+import { useState, useEffect, useRef } from 'react';
+
+const PROMPTS = [
+    "현재 감지되는 '외부 인력(그리움)'의 강도는 어떠하며, 이에 대응하기 위해 나는 어떤 '코어 활동'을 하고 있나요?",
+    "시스템(나)의 안정성을 높이기 위해, 지금 당장 투입할 수 있는 긍정적 데이터(행동)는 무엇인가요?",
+    "외부 신호(연락하고 싶은 충동)를 훌륭하게 제어한 나 자신에게, 관리자로서 격려의 로그를 남겨주세요.",
+    "관계 변수를 제외하고, 오늘 나의 '성장 알고리즘'에서 가장 잘 작동한 부분은 무엇인가요?",
+    "현재 나의 물리적 리소스(체력, 허기) 상태는 어떠한가요? 재충전이 필요한가요?",
+    "미래에 대한 불확실한 예측 프로세스를 끄고, '오늘'이라는 런타임에만 집중한다면 무엇을 실행해야 할까요?"
+];
+
+/**
+ * useInnerOrbit - Inner Orbit 애플리케이션의 상태 및 비즈니스 로직을 관리하는 커스텀 훅
+ * 
+ * @returns {Object} 상태, 액션, ref를 포함하는 객체
+ */
+export default function useInnerOrbit() {
+    // ===== 상태 관리 =====
+    const [entries, setEntries] = useState(() => {
+        try {
+            const saved = localStorage.getItem('journalEntries');
+            return saved ? JSON.parse(saved) : [];
+        } catch (e) {
+            return [];
+        }
+    });
+
+    const [message, setMessage] = useState("");
+    const [stability, setStability] = useState(50);
+    const [gravity, setGravity] = useState(50);
+    const [view, setView] = useState('write');
+    const [promptIndex, setPromptIndex] = useState(0);
+    const [showPrompt, setShowPrompt] = useState(true);
+    const [isHeaderExpanded, setIsHeaderExpanded] = useState(true);
+    const fileInputRef = useRef(null);
+
+    // ===== localStorage 영속성 관리 =====
+    useEffect(() => {
+        localStorage.setItem('journalEntries', JSON.stringify(entries));
+    }, [entries]);
+
+    // ===== CRUD 작업 =====
+
+    /**
+     * 새 로그 엔트리를 생성하고 저장
+     */
+    const handleSubmit = () => {
+        if (!message.trim()) return;
+
+        const newEntry = {
+            id: Date.now(),
+            date: new Date().toLocaleString('ko-KR', {
+                year: 'numeric', month: 'long', day: 'numeric',
+                hour: '2-digit', minute: '2-digit', hour12: false
+            }),
+            content: message,
+            stability,
+            gravity,
+        };
+
+        setEntries([newEntry, ...entries]);
+        setMessage("");
+        setView('history');
+    };
+
+    /**
+     * 로그 엔트리 삭제
+     * @param {number} id - 삭제할 엔트리의 ID
+     */
+    const deleteEntry = (id) => {
+        if (window.confirm('이 로그를 영구 삭제하시겠습니까?')) {
+            setEntries(entries.filter(e => e.id !== id));
+        }
+    };
+
+    /**
+     * 현재 프롬프트를 메시지 영역에 삽입
+     */
+    const insertPrompt = () => {
+        const promptText = PROMPTS[promptIndex];
+        setMessage((prev) => prev ? `${prev}\n\n[Query]: ${promptText}\n[Log]: ` : `[Query]: ${promptText}\n[Log]: `);
+    };
+
+    // ===== 데이터 백업/복원 =====
+
+    /**
+     * 모든 엔트리를 JSON 파일로 다운로드
+     */
+    const downloadData = () => {
+        const dataStr = JSON.stringify(entries, null, 2);
+        const blob = new Blob([dataStr], { type: "application/json" });
+        const url = URL.createObjectURL(blob);
+        const link = document.createElement('a');
+        link.href = url;
+        link.download = `inner_orbit_backup_${new Date().toISOString().slice(0, 10)}.json`;
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+    };
+
+    /**
+     * JSON 파일에서 데이터를 불러와 병합
+     * @param {Event} event - 파일 입력 이벤트
+     */
+    const handleFileUpload = (event) => {
+        const file = event.target.files[0];
+        if (!file) return;
+        const reader = new FileReader();
+        reader.onload = (e) => {
+            try {
+                const imported = JSON.parse(e.target.result);
+                if (Array.isArray(imported)) {
+                    if (window.confirm(`${imported.length}개의 데이터 로그를 병합하시겠습니까?`)) {
+                        const existingIds = new Set(entries.map(e => e.id));
+                        const newUnique = imported.filter(e => !existingIds.has(e.id));
+                        setEntries([...newUnique, ...entries].sort((a, b) => b.id - a.id));
+                    }
+                }
+            } catch (error) { alert('데이터 파싱 오류가 발생했습니다.'); }
+        };
+        reader.readAsText(file);
+        event.target.value = '';
+    };
+
+    // ===== 반환 인터페이스 =====
+    return {
+        // 상태
+        entries,
+        gravity,
+        stability,
+        message,
+        view,
+        promptIndex,
+        showPrompt,
+        isHeaderExpanded,
+        prompts: PROMPTS,
+
+        // Setters
+        setMessage,
+        setStability,
+        setGravity,
+        setView,
+        setPromptIndex,
+        setShowPrompt,
+        setIsHeaderExpanded,
+
+        // Actions
+        handleSubmit,
+        deleteEntry,
+        insertPrompt,
+        downloadData,
+        handleFileUpload,
+
+        // Refs
+        fileInputRef
+    };
+}
